Close #215 - Rename extras-hedgehog-cats-effect3 to extras-hedgehog-ce3
