+++ modules/extras-hedgehog-ce3/shared/src/test/scala/extras/hedgehog/ce3/CatsEffectRunnerSpec.scala
@@ -1,4 +1,4 @@
-package extras.hedgehog.cats.effect
+package extras.hedgehog.ce3
 
 import cats.{Eq, Show}
 import cats.effect.IO
